+++ app/lib/ui/components/list_item/list_item_code_generator.dart
@@ -58,7 +58,7 @@ class ListItemCodeGenerator {
       if (!state.enable) '  enable: false,',
       if (state.clickable) '  onTap: () { /* navigate */ },',
       if (state.clickable) '  indicator: ${_indicatorCode(state.indicator)},',
-      if (!state.edgeToEdge) '  edgeToEdge: false,',
+      '  edgeToEdge: ${state.edgeToEdge},',
     ];
     return 'OudsListItem(\n${params.join('\n')}\n)';
   }
@@ -84,7 +84,7 @@ class ListItemCodeGenerator {
       if (!state.enable) '  enable: false,',
       if (state.clickable) '  onTap: () { /* navigate */ },',
       if (state.clickable) '  indicator: ${_indicatorCode(state.indicator)},',
-      if (!state.edgeToEdge) '  edgeToEdge: false,',
+      '  edgeToEdge: ${state.edgeToEdge},',
     ];
     return 'OudsSmallListItem(\n${params.join('\n')}\n)';
   }