+++ app/lib/ui/utilities/customizable/customizable_widget_state.dart
@@ -23,6 +23,7 @@ abstract class CustomizationWidgetState<T extends StatefulWidget>
   late final TextState textState;
   late final SelectState selectState;
   late final CenterAlignedState centerAlignedState;
+  late final EdgeState edgeToEdgeState;
   late final TintedState tintedState;
 
   @override
@@ -35,6 +36,7 @@ abstract class CustomizationWidgetState<T extends StatefulWidget>
     selectState = SelectState(setState);
     //needed for top bars
     centerAlignedState = CenterAlignedState(setState);
+    edgeToEdgeState = EdgeState(setState, initial: true);
     tintedState = TintedState(setState);
   }
 
@@ -57,6 +59,9 @@ abstract class CustomizationWidgetState<T extends StatefulWidget>
 
   bool get hasCentredAligned => centerAlignedState.value;
   set hasCentredAligned(bool value) => centerAlignedState.value = value;
+
+  bool get edgeToEdge => edgeToEdgeState.value;
+  set edgeToEdge(bool value) => edgeToEdgeState.value = value;
 
   bool get isTinted => tintedState.selected == TintedEnum.tinted;
 
@@ -150,6 +155,21 @@ class CenterAlignedState {
   set value(bool newValue) {
     _setState(() {
       _hasCentredAligned = newValue;
+    });
+  }
+}
+
+/// Edge to edge state management
+class EdgeState {
+  EdgeState(this._setState, {bool initial = false}) : _value = initial;
+
+  final void Function(void Function()) _setState;
+  bool _value;
+
+  bool get value => _value;
+  set value(bool newValue) {
+    _setState(() {
+      _value = newValue;
     });
   }
 }