+++ ouds_core/lib/components/checkbox/ouds_checkbox_item.dart
@@ -56,6 +56,8 @@ import 'package:ouds_core/components/control/ouds_control_item_icon.dart';
 ///   Defaults to `false`.
 /// - [errorText]: Text shown below the checkbox item indicating an error state. Supports only strong text formatting using `**bold**`.
 ///   Rich text is supported only for error messages.
+/// - [edgeToEdge] Controls the horizontal layout of the checkbox item. When `true`, the checkbox item is designed to span the full width of the screen or container. When `false`,
+///   it is adapted for use within constrained layouts or containers with their own padding. Defaults to `true`.
 ///
 ///
 /// ### You can use [OudsCheckboxItem] component in your project, customizing parameters as needed :
@@ -76,6 +78,7 @@ import 'package:ouds_core/components/control/ouds_control_item_icon.dart';
 ///   icon: OudsControlItemIcon('assets/ic_heart/svg'),
 ///   isError: false,
 ///   divider: true,
+///   edgeToEdge: true,
 ///   onChanged: (bool? value) {
 ///      setState(() {
 ///         isChecked = value;
@@ -100,6 +103,7 @@ class OudsCheckboxItem extends StatelessWidget {
   final bool divider;
   final bool tristate;
   final bool constrainedMaxWidth;
+  final bool edgeToEdge;
 
   const OudsCheckboxItem({
     super.key,
@@ -116,6 +120,7 @@ class OudsCheckboxItem extends StatelessWidget {
     this.divider = false,
     this.tristate = false,
     this.constrainedMaxWidth = false,
+    this.edgeToEdge = true,
   });
 
   @override
@@ -143,6 +148,7 @@ class OudsCheckboxItem extends StatelessWidget {
       divider: divider,
       constrainedMaxWidth: constrainedMaxWidth,
       reversed: reversed,
+      edgeToEdge: edgeToEdge,
       onTap: onChanged != null
           ? () {
               bool? newValue;